Update the Cards section to contain separate classes for Hand and Deck.
