+++ src/BiddingFacility/BiddingFacilityDriver.cpp
@@ -2,6 +2,7 @@
 #include <iostream>
 #include <string>
 #include <list>
+#include <vector>
 
 using namespace std;
 
@@ -11,19 +12,21 @@ using namespace std;
 // 2. Show bids which are tied
 // 3. Show if all bids are zero
 
-void DemonstrateBiddingFacility(const int numberOfPlayers){
+void DemonstrateBiddingFacility(int numberOfPlayers){
 
     int whiteLines = 100;
 
     cout << endl << "Beginning Bidding Phase!" << endl << endl;
 
     // Initialization of BiddingFacilities, each player must enter their last name and starting bid amount
-    BiddingFacility* biddingFacilities[numberOfPlayers];
+    vector<BiddingFacility> biddingFacilities;
     for(int i = 0; i < numberOfPlayers; i++){
         cout << "Bidding details for player #" << (i + 1) << endl;
-        biddingFacilities[i] = new BiddingFacility;
-        biddingFacilities[i]->ReceiveStartingCoins(numberOfPlayers);
-        biddingFacilities[i]->EnterBid();
+        BiddingFacility newBiddingFacility;
+        newBiddingFacility.ReceiveStartingCoins(numberOfPlayers);
+        newBiddingFacility.EnterBid();
+        biddingFacilities.push_back(newBiddingFacility);
+
         for(int j = 0; j < whiteLines; j++){
             cout << endl;
         }
@@ -31,35 +34,33 @@ void DemonstrateBiddingFacility(const int numberOfPlayers){
 
     // Each player reveals their bid
     for(int i = 0; i < numberOfPlayers; i++) {
-        biddingFacilities[i]->RevealBid();
+        biddingFacilities[i].RevealBid();
     }
 
     // Determine which player(s) placed the highest bid
-    BiddingFacility* bidWinner;
+    BiddingFacility *bidWinner;
     int highestBid = -1;
     for(int i = 0; i < numberOfPlayers; i++){
-        if(*(biddingFacilities[i]->GetBidAmount()) > highestBid) {
-            highestBid = *(biddingFacilities[i]->GetBidAmount());
-            bidWinner = biddingFacilities[i];
+        if(*(biddingFacilities[i].GetBidAmount()) > highestBid) {
+            highestBid = *(biddingFacilities[i].GetBidAmount());
+            bidWinner = &biddingFacilities[i];
         }
     }
     for(int i = 0; i < numberOfPlayers; i++){
-        if(*(biddingFacilities[i]->GetBidAmount()) == highestBid &&
-            (biddingFacilities[i]->GetLastName())->compare(*(bidWinner->GetLastName())) < 0){
-            bidWinner = biddingFacilities[i];
+        if(*(biddingFacilities[i].GetBidAmount()) == highestBid &&
+            (biddingFacilities[i].GetLastName())->compare(*(bidWinner->GetLastName())) < 0){
+            bidWinner = &biddingFacilities[i];
         }
     }
     for(int i = 0; i < numberOfPlayers; i++){
-        if(biddingFacilities[i] == bidWinner){
-            biddingFacilities[i]->ResolveBid(true);
+        if(&biddingFacilities[i] == bidWinner){
+            biddingFacilities[i].ResolveBid(true);
         }else{
-            biddingFacilities[i]->ResolveBid(false);
+            biddingFacilities[i].ResolveBid(false);
         }
     }
 
     // Memory clean up
-    for(int i = 0; i < numberOfPlayers; i++){
-        delete biddingFacilities[i];
-        biddingFacilities[i] = nullptr;
-    }
+    delete bidWinner;
+    bidWinner = nullptr;
 }
