+++ src/Cards/Cards.h
@@ -14,8 +14,6 @@ public:
     string* GetGoodsAndAction();
     int* GetCost();
     void SetCost(int _cost);
-    static void Draw(vector<Cards> *deckPtr, vector<Cards> *faceUpCardsPtr);
-    static void ShuffleDeck(vector <Cards> deck);
     Cards();
     ~Cards();
     Cards(const Cards &copy);
@@ -27,29 +25,3 @@ public:
     friend istream & operator >> (istream &in, Cards &c);
 };
 
-class Deck{
-private:
-    static vector<Cards>* deck;
-    static vector<Cards>* faceUpCards;
-public:
-    Deck();
-    Deck(int numberOfPlayers);
-    ~Deck();
-    static vector<Cards>* GetDeck();
-    static vector<Cards>* GetFaceUpCards();
-    void PrintCardsIn(vector<Cards>* vectorOfCards);
-    void PopulateFaceUpCards();
-    void SetFaceUpCardsCost();
-    //void ShuffleDeck();
-    void Draw();
-};
-
-class Hand{
-private:
-    vector<Cards>* hand;
-public:
-    Hand();
-    ~Hand();
-    void AddToHand(Cards card);
-    void Exchange(int positionToPickUp);
-};
